requirements.txt
PyQt5
mysql-connector-python
pandas
matplotlib
fpdf
